Explicit int casts

+++ stitch.py
@@ -26,8 +26,8 @@ def stitch(metadata, ims):
     print("Making a {}x{} image.".format(W, H))
     canvas = np.zeros((H, W), np.uint16)
 
-    h_margin = np.rint(0.1*h)
-    w_margin = np.rint(0.1*w)
+    h_margin = np.rint(0.1*h).astype(int)
+    w_margin = np.rint(0.1*w).astype(int)
     fade = np.ones((h, w), np.double)
     h_ascending = np.linspace(0, 1, h_margin+1, endpoint=False)[1:].reshape(-1, 1)
     fade[:h_margin,:] *= h_ascending
@@ -42,9 +42,9 @@ def stitch(metadata, ims):
             print i,
             sys.stdout.flush()
         r_n, c_n = grid_index[i]
-        r_px = (N_row - 1 - r_n) * (0.9 * h)
-        c_px = (N_col - 1 - c_n) * (0.9 * w)
-        canvas[r_px:r_px+h, c_px:c_px+w] += np.rint(ims[i] * fade)
+        r_px = int((N_row - 1 - r_n) * (0.9 * h))
+        c_px = int((N_col - 1 - c_n) * (0.9 * w))
+        canvas[r_px:r_px+h, c_px:c_px+w] += np.rint(ims[i] * fade).astype(np.uint16)
     print
     return canvas
